refactor(wallet): remove unused modules and consolidate token standards
