+++ .gitignore
@@ -5,3 +5,7 @@
 
 # Node
 node_modules
+
+# .env files containing secrets or sensitive data
+.env
+.env.local
