+++ canisters/rest-api/src/app.module.ts
@@ -1,13 +1,11 @@
 import { Module } from '@nestjs/common';
 
 import { CoreModule } from './core/core.module';
-import { NetworkModule } from './network/network.module';
-import { TokenModule } from './token/token.module';
 import { UserModule } from './user/user.module';
 import { WalletModule } from './wallet/wallet.module';
 
 @Module({
-  imports: [CoreModule, UserModule, WalletModule, TokenModule, NetworkModule],
+  imports: [CoreModule, UserModule, WalletModule],
   controllers: [],
   providers: [],
 })
