--- canisters/rest-api/src/network/dtos/register-network.dto.ts
@@ -1,6 +1,0 @@
-import { IsString } from 'class-validator';
-
-export class RegisterNetworkDto {
-  @IsString()
-  name: string;
-}
