--- canisters/rest-api/src/network/entities/network.entity.ts
@@ -1,7 +1,0 @@
-import { Record, text } from 'azle/experimental';
-
-export const Network = Record({
-  name: text,
-});
-
-export type Network = typeof Network.tsType;
