--- canisters/rest-api/src/network/network.module.ts
@@ -1,11 +1,0 @@
-import { Module } from '@nestjs/common';
-
-import { NetworkController } from './network.controller';
-import { NetworkService } from './network.service';
-
-@Module({
-  controllers: [NetworkController],
-  providers: [NetworkService],
-  exports: [NetworkService],
-})
-export class NetworkModule {}
