--- canisters/rest-api/src/token/dtos/register-token.dto.ts
@@ -1,14 +1,0 @@
-import { ArrayNotEmpty, IsArray, IsString } from 'class-validator';
-
-export class RegisterTokenDto {
-  @IsString()
-  name: string;
-
-  @IsString()
-  symbol: string;
-
-  @IsArray()
-  @ArrayNotEmpty()
-  @IsString({ each: true })
-  networks: string[];
-}
