--- canisters/rest-api/src/token/entities/token.entity.ts
@@ -1,9 +1,0 @@
-import { Record, Vec, text } from 'azle/experimental';
-
-export const Token = Record({
-  name: text,
-  symbol: text,
-  networks: Vec(text),
-});
-
-export type Token = typeof Token.tsType;
